requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-Login==0.6.3
gunicorn==21.2.0
werkzeug==3.0.0
openpyxl==3.1.2
psycopg2-binary==2.9.9
